refactor(cosmo_val): extract CatalogCharacterizationMixin (catalog_characterization)
